requirements.txt
SpeechRecognition
PyAudio
google-cloud-speech
google-cloud-texttospeech
BanglaTTS
google-cloud-aiplatform
fastapi
uvicorn
python-multipart